Remove ChartRenderer and rename chartItems2 to chartItems.
