+++ lib/ReactViews/Custom/Chart/Chart.jsx
@@ -1,301 +1,168 @@
-"use strict";
-
-// For documentation on the custom <chart> tag, see lib/Models/registerCustomComponentTypes.js.
-//
-// Two possible approaches to combining D3 and React:
-// 1. Render SVG element in React, let React keep control of the DOM.
-// 2. React treats the element like a blackbox, and D3 is in control.
-// We take the second approach, because it gives us much more of the power of D3 (animations etc).
-//
-// See also:
-// https://facebook.github.io/react/docs/working-with-the-browser.html
-// http://ahmadchatha.com/writings/article1.html
-// http://nicolashery.com/integrating-d3js-visualizations-in-a-react-app/
-
-import React from "react";
-
-import PropTypes from "prop-types";
-import createReactClass from "create-react-class";
-
-import defined from "terriajs-cesium/Source/Core/defined";
-import defaultValue from "terriajs-cesium/Source/Core/defaultValue";
-import DeveloperError from "terriajs-cesium/Source/Core/DeveloperError";
-// import loadText from "../../../Core/loadText";
-// import when from "terriajs-cesium/Source/ThirdParty/when";
+import uniqBy from "lodash/uniqBy";
+import { action, computed, observable } from "mobx";
 import { observer } from "mobx-react";
-// import ChartData from "../../../Charts/ChartData";
-import ChartRenderer from "../../../Charts/ChartRenderer";
-// import proxyCatalogItemUrl from "../../../Models/proxyCatalogItemUrl";
-import CsvCatalogItem from "../../../Models/CsvCatalogItem";
+import PropTypes from "prop-types";
+import React from "react";
+import {
+  VictoryAxis,
+  VictoryChart,
+  VictoryLegend,
+  VictoryLine,
+  VictoryTheme
+} from "victory";
 
-import Styles from "./chart.scss";
-
-const defaultHeight = 100;
-// const defaultColor = undefined; // Allows the line color to be set by the css, esp. in the feature info panel.
-
-const Chart = observer(
-  createReactClass({
-    displayName: "Chart",
-
-    // this._element is updated by the ref callback attribute, https://facebook.github.io/react/docs/more-about-refs.html
-    _element: undefined,
-
-    _promise: undefined,
-
-    _tooltipId: undefined,
-
-    propTypes: {
-      domain: PropTypes.object,
-      // A presentation mode, one of:
-      //   "feature-info": makes a "mini-chart" with no grid, less space, for use in a feature info window
-      //   "histogram": a bit less space
-      //   undefined: default styling
-      styling: PropTypes.string, // nothing, 'feature-info' or 'histogram' -- TODO: improve
-      height: PropTypes.number,
-      axisLabel: PropTypes.object,
-      catalogItem: PropTypes.object,
-      transitionDuration: PropTypes.number,
-      highlightX: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
-      updateCounter: PropTypes.any, // Change this to trigger an update.
-      pollSeconds: PropTypes.any, // This is not used by Chart. It is used internally by registerCustomComponentTypes.
-
-      // You can provide the data directly via props.data (ChartData[]):
-      data: PropTypes.array,
-      // chartType: PropTypes.object, // TODO clarify. ChartData has its own 'type' which can be bar, line, etc.
-
-      // Or, provide a URL to the data, along with optional xColumn, yColumns, colors
-      url: PropTypes.string,
-      xColumn: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
-      yColumns: PropTypes.array,
-      colors: PropTypes.array,
-      pollUrl: PropTypes.string,
-
-      // Or, provide a tableStructure directly.
-      tableStructure: PropTypes.object,
-
-      items: PropTypes.array
-    },
-
-    chartDataArrayFromTableStructure(table) {
-      return table.chartItems;
-    },
-
-    getChartDataPromise(data, url, catalogItem) {
-      // Returns a promise that resolves to an array of ChartData.
-      const that = this;
-      if (defined(data)) {
-        // Nothing to do - the data was provided (either as props.data or props.tableStructure).
-        const promises = data.map(item => item.loadMapItems());
-        return Promise.all(promises).then(items => {
-          return data.reduce((p, c) => p.concat(c.chartItems), []);
-        });
-      } else if (defined(url)) {
-        return loadIntoTableStructure(catalogItem, url)
-          .then(that.chartDataArrayFromTableStructure)
-          .catch(function(e) {
-            // It looks better to create a blank chart than no chart.
-            return [];
-          });
-      }
-    },
-
-    componentDidMount() {
-      const that = this;
-      const chartParameters = that.getChartParameters();
-      const promise = that.getChartDataPromise(
-        chartParameters.data,
-        that.props.url,
-        that.props.catalogItem
-      );
-      promise.then(function(chartItems) {
-        chartParameters.chartItems = chartItems;
-        ChartRenderer.create(that._element, chartParameters);
-      });
-      that._promise = promise.then(function() {
-        // that.rnd = Math.random();
-        // React should handle the binding for you, but it doesn't seem to work here; perhaps because it is inside a Promise?
-        // So we return the bound listener function from the promise.
-        const resize = function() {
-          // This function basically the same as componentDidUpdate, but it speeds up transitions.
-          // Note same caveats - doesn't work if the data came from a URL.
-          // That's ok for our purposes, since the URL is only used in a feature info panel, which never resizes dynamically.
-          if (that._element) {
-            const localChartParameters = that.getChartParameters();
-            if (defined(chartParameters.data)) {
-              localChartParameters.transitionDuration = 1;
-              ChartRenderer.update(that._element, localChartParameters);
-            }
-          } else {
-            // This would happen if event listeners were not properly removed (ie. if you get this error, a bug was introduced to this code).
-            throw new DeveloperError("Missing chart DOM element " + that.url);
-          }
-        };
-        // console.log('Listening for resize on', that.props.url, that.rnd, boundComponentDidUpdate);
-        window.addEventListener("resize", resize);
-        return resize;
-      });
-    },
-
-    componentDidUpdate() {
-      // Update the chart with props.data or props.tableStructure, if present.
-      // If the data came from a URL, there are three possibilities:
-      // 1. The URL has changed.
-      // 2. The URL is the same and therefore we do not want to reload it.
-      // 3. The URL is the same, but the chart came from a self-updating <chart> tag (ie. one with a poll-seconds attribute),
-      //    and so we do want to reload it.
-      // Note that registerCustomComponent types wraps its charts in a div with a key based on the url,
-      // so if the URL has changed, it actually mounts a new component, thereby triggering a load.
-      // (Ie. we don't need to cover case (1) here.)
-      // In case (3), props.updateCounter will be set to an integer, and we should update the data from the URL.
-      const element = this._element;
-      const chartParameters = this.getChartParameters();
-      if (defined(chartParameters.data)) {
-        ChartRenderer.update(element, chartParameters);
-      } else if (this.props.updateCounter > 0) {
-        // The risk here is if it's a time-varying csv with <chart> polling as well.
-        const url = this.props.pollUrl || this.props.url;
-        const promise = this.getChartDataPromise(
-          chartParameters.data,
-          url,
-          this.props.catalogItem
-        );
-        promise.then(function(chartItems) {
-          chartParameters.chartItems = chartItems;
-          ChartRenderer.update(element, chartParameters);
-        });
-      }
-      ChartRenderer.update(element, chartParameters);
-    },
-
-    componentWillUnmount() {
-      const that = this;
-      this._promise.then(function(listener) {
-        window.removeEventListener("resize", listener);
-        // console.log('Removed resize listener for', that.props.url, that.rnd, listener);
-        ChartRenderer.destroy(that._element, that.getChartParameters());
-        that._element = undefined;
-      });
-      this._promise = undefined;
-    },
-
-    /**
-     * Return the initialisation parameters to be passed to LineChart (or other chart type).
-     * If it is not a mini-chart, add tooltip settings (including a unique id for the tooltip DOM element).
-     */
-    getChartParameters() {
-      let margin;
-      let tooltipSettings;
-      let titleSettings;
-      let grid;
-      if (this.props.styling !== "feature-info") {
-        if (!defined(this._tooltipId)) {
-          // In case there are multiple charts with tooltips. Unlikely to pick the same random number. Remove the initial "0.".
-          this._tooltipId =
-            "d3-tooltip-" +
-            Math.random()
-              .toString()
-              .substr(2);
-        }
-        tooltipSettings = {
-          className: Styles.toolTip,
-          id: this._tooltipId,
-          align: "prefer-right", // With right/left alignment, the offset is relative to the svg, so need to inset.
-          offset: { top: 40, left: 33, right: 30, bottom: 5 }
-        };
-        if (this.props.styling === "histogram") {
-          titleSettings = undefined;
-          margin = { top: 0, right: 0, bottom: 0, left: 0 };
-        } else {
-          margin = {
-            top: 8,
-            right: 16,
-            bottom: 20,
-            left: 0
-          };
-
-          titleSettings = {
-            type: "legend",
-            height: 30
-          };
-        }
-        grid = {
-          x: true,
-          y: true
-        };
-      }
-      if (defined(this.props.highlightX)) {
-        tooltipSettings = undefined;
-      }
-
-      let chartData;
-      if (defined(this.props.data)) {
-        chartData = this.props.data;
-      } else if (defined(this.props.tableStructure)) {
-        chartData = this.chartDataArrayFromTableStructure(
-          this.props.tableStructure
-        );
-      } else if (defined(this.props.items)) {
-        chartData = this.props.items;
-      }
-
-      // for better presentation, we order datasets  so that the ones with units information to
-      // display first, so that Yaxis with unit shows up outside yaxis without
-      // unit
-      if (defined(chartData) && chartData.length > 1) {
-        chartData = chartData
-          .slice()
-          .sort((data1, data2) => defined(data1.units) - defined(data2.units));
-      }
-
-      return {
-        data: chartData,
-        chartItems: chartData.reduce((p, c) => p.concat(c.chartItems), []),
-        domain: this.props.domain,
-        width: "100%",
-        height: defaultValue(this.props.height, defaultHeight),
-        axisLabel: this.props.axisLabel,
-        mini: this.props.styling === "feature-info",
-        transitionDuration: this.props.transitionDuration,
-        margin: margin,
-        tooltipSettings: tooltipSettings,
-        titleSettings: titleSettings,
-        grid: grid,
-        highlightX: this.props.highlightX
-      };
-    },
-
-    render() {
-      return (
-        <div className={Styles.chart}>
-          <div
-            className={Styles.chartInner}
-            ref={element => {
-              this._element = element;
-            }}
-          />
-        </div>
-      );
-    }
-  })
-);
+const chartMinWidth = 110; // Required to prevent https://github.com/FormidableLabs/victory-native/issues/132
 
 /**
- * Loads data from a URL into a table structure.
- * @param  {String} url The URL.
- * @return {Promise} A promise which resolves to a table structure.
+ * A chart component that implements the charting behavior common to
+ * FeatureInfoPanelChart & BottomDockChart. Presentation aspects are further
+ * customized in the individual components.
+ *
  */
-function loadIntoTableStructure(catalogItem, url) {
-  const result = new CsvCatalogItem(undefined, catalogItem.terria);
-  result.setTrait("definition", "url", url);
-  const chartStyle = result.addObject("definition", "styles", "chart");
-  chartStyle.chart.setTrait("definition", "xAxisColumn", "Time (UTC)");
-  chartStyle.chart.setTrait("definition", "yAxisColumn", "MW");
-  result.setTrait("definition", "activeStyle", "chart");
+@observer
+class Chart extends React.Component {
+  static propTypes = {
+    width: PropTypes.number,
+    height: PropTypes.number,
+    chartItems: PropTypes.array.isRequired,
+    xAxis: PropTypes.object.isRequired,
+    containerComponent: PropTypes.element,
+    theme: PropTypes.object,
+    renderLegends: PropTypes.func,
+    renderXAxis: PropTypes.func,
+    renderYAxis: PropTypes.func,
+    renderLine: PropTypes.func
+  };
 
-  // TODO: should be loadChartItems
-  return result.loadMapItems().then(() => {
-    return result;
-  });
+  static defaultProps = {
+    height: 110,
+    theme: VictoryTheme.material,
+    renderXAxis: label => <VictoryAxis label={label} />,
+    renderYAxis: ({ label }, i) => (
+      <VictoryAxis dependentAxis key={i} label={label} />
+    ),
+    renderLegends: (legends, width) => (
+      <VictoryLegend x={width / 2} orientation="horizontal" data={legends} />
+    )
+  };
+
+  @computed
+  get yAxes() {
+    return uniqBy(this.props.chartItems, ({ units }) => units).map(data => ({
+      units: data.units,
+      color: data.getColor()
+    }));
+  }
+
+  @computed
+  get legends() {
+    return this.props.chartItems.map(({ name, getColor }) => {
+      return { name, symbol: { fill: getColor() } };
+    });
+  }
+
+  renderLine(data, index) {
+    return (
+      <VictoryLine
+        key={index}
+        data={data.points.map(p => ({
+          ...p,
+          units: data.units,
+          name: data.name
+        }))}
+        style={{
+          data: { stroke: data.getColor() },
+          labels: { fill: data.getColor() }
+        }}
+      />
+    );
+  }
+
+  renderChartItem(chartItem, index) {
+    // TODO: render data based on data.type
+    const renderLine = this.props.renderLine || this.renderLine;
+    return renderLine(chartItem, index);
+  }
+
+  renderChart(width, height) {
+    if (!this.props.xAxis) {
+      return null;
+    }
+
+    return (
+      <VictoryChart
+        width={width}
+        height={height}
+        theme={this.props.theme}
+        scale={{ x: this.props.xAxis.scale }}
+        containerComponent={this.props.containerComponent}
+      >
+        {this.props.renderLegends(this.legends, width)}
+        {this.props.renderXAxis(this.props.xAxis.units)}
+        <For each="yAxis" index="i" of={this.yAxes}>
+          {this.props.renderYAxis(yAxis, i, this.yAxes.length)}
+        </For>
+        <For each="chartItem" index="i" of={this.props.chartItems}>
+          {this.renderChartItem(chartItem, i)}
+        </For>
+      </VictoryChart>
+    );
+  }
+
+  render() {
+    return (
+      <Sized>
+        {({ width: parentWidth }) =>
+          this.renderChart(
+            Math.max(chartMinWidth, this.props.width || parentWidth),
+            this.props.height
+          )
+        }
+      </Sized>
+    );
+  }
+}
+
+/**
+ * A component that passes its width to its child component.
+ */
+@observer
+class Sized extends React.Component {
+  @observable containerElement = undefined;
+  @observable width = 0;
+
+  static propTypes = {
+    children: PropTypes.func.isRequired
+  };
+
+  @action
+  attachElement(el) {
+    this.containerElement = el;
+  }
+
+  componentDidMount() {
+    this.updateWidth();
+    window.addEventListener("resize", this.updateWidth.bind(this));
+  }
+
+  componentWillUnmount() {
+    window.removeEventListener("resize", this.updateWidth.bind(this));
+  }
+
+  @action
+  updateWidth() {
+    if (this.containerElement) {
+      this.width = this.containerElement.getBoundingClientRect().width;
+    }
+  }
+
+  render() {
+    return (
+      <div ref={this.attachElement.bind(this)}>
+        {this.props.children({ width: this.width })}
+      </div>
+    );
+  }
 }
 
 export default Chart;
